requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.12
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
websockets>=14.0
python-dotenv>=1.0.1
aiofiles>=23.2.0
cryptography>=41.0.0
pillow>=10.0.0

